Made changes to some of the views as needed to make the frontend work the way I wanted it, also finished all functionality for every page adding the ability to add runs/races/goals while also being able to delete added runs and races.

+++ back_end/user_app/models.py
@@ -8,7 +8,7 @@ class App_user(AbstractUser):
     state = models.CharField(max_length=2)
     age = models.IntegerField(validators=[v.MinValueValidator(13), v.MaxValueValidator(120)])
     display_name = models.CharField(unique=True)
-    total_distance = models.IntegerField(default=0)
-    total_time = models.IntegerField(default=0)
+    total_distance = models.DecimalField(max_digits=100, decimal_places=2,default=0)
+    total_time = models.DecimalField(max_digits=100, decimal_places=2,default=0)
     USERNAME_FIELD = 'email'
     REQUIRED_FIELDS =[state]
